requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

# ORM & DB
sqlalchemy==2.0.30

# Password hashing
passlib[bcrypt]==1.7.4

# JWT support
python-jose==3.3.0

# Pydantic (comes with FastAPI, but version can be pinned if needed)
pydantic==2.7.1

pydantic[email]==2.7.1

bcrypt>=4.0.1
passlib>=1.7.4
